refactor: move PR response instructions to dedicated file and bump version
