+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cr-mcp",
-  "version": "1.0.11",
+  "version": "1.0.12",
   "description": "Code Review Model Context Protocol (CR-MCP) server for managing GitHub pull request comments",
   "main": "dist/server.js",
   "types": "dist/server.d.ts",
